Se agrega reduccionCupo() + Test a Atraccion

+++ src/turismotierramedia/Producto.java
@@ -14,11 +14,12 @@ public class Producto {
 		this.tipoAtraccion = tipoAtraccion;
 	}
 	
-	public Producto(String nombre, int costo, int tiempo) {
+	public Producto(String nombre, int costo, double tiempo) {
 		this.nombre = nombre;
 		this.costo = costo;
 		this.tiempo = tiempo;
 	}
+
 
 	protected int getCosto() {
 		return costo;
